requirements.txt
fastapi
uvicorn[standard]
jinja2
numpy
sentence-transformers
chromadb
requests
python-dotenv
